+++ admin/src/View/Titlesets/HtmlView.php
@@ -10,6 +10,7 @@ use Joomla\CMS\Language\Text;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
 use Joomla\CMS\Router\Route;
 use Joomla\CMS\Toolbar\ToolbarHelper;
+use Joomla\CMS\Toolbar\Toolbar;
 use CB\Component\Contentbuilderng\Administrator\Model\TitlesetsModel;
 use Joomla\CMS\Application\AdministratorApplication;
 use Joomla\CMS\Factory;
@@ -35,16 +36,18 @@ final class HtmlView extends BaseHtmlView
         $this->items = $model->getItems();
 
         ToolbarHelper::title(Text::_('COM_CONTENTBUILDERNG_TITLESETS_TITLE'), 'list');
-        ToolbarHelper::link(
-            Route::_('index.php?option=com_contentbuilderng&view=titleset', false),
-            'JTOOLBAR_NEW',
-            'new'
-        );
-        ToolbarHelper::link(
-            Route::_('index.php?option=com_contentbuilderng&view=about', false),
-            'JTOOLBAR_CLOSE',
-            'cancel'
-        );
+        /** @var Toolbar $toolbar */
+        $toolbar = $this->getDocument()->getToolbar('toolbar');
+        $toolbar->linkButton('titleset-new')
+            ->url(Route::_('index.php?option=com_contentbuilderng&view=titleset', false))
+            ->text('JTOOLBAR_NEW')
+            ->icon('icon-new')
+            ->attributes(['title' => Text::_('COM_CONTENTBUILDERNG_TITLESETS_NEW_DESC')]);
+        $toolbar->linkButton('titlesets-close')
+            ->url(Route::_('index.php?option=com_contentbuilderng&view=about', false))
+            ->text('JTOOLBAR_CLOSE')
+            ->icon('icon-cancel')
+            ->attributes(['title' => Text::_('COM_CONTENTBUILDERNG_TITLESETS_CLOSE_DESC')]);
 
         parent::display($tpl);
     }